In param_calibration.py, modified code for determining simulation time points in `create_figures()` to explicitly include experimental time points. Need this in case we need to scale the results by the result at a particular expt time point. Also, added code to `plot_timecourses()` to add hatch patterns to the simulated envelopes if there are multiple experiments for a given `expt_id` (i.e., using the `ParallelExperiments` class) and to change the marker for plotting the expt data points.
In sim_protocols.py, changed the way the scaling is handled in the `ParallelExperiments` class. Now, the user passes the observable name together with the experimental index and the time value (not the index) that the results should be scaled by. This is passed in as a dictionary. A new method called `find_exact_index` has been added to locate the index of the time point in the `tspan` array, taking into account possible numerical precision issues. With this, we can now run simulations with more time points than the experimental time points and still scale the results at the correct time point to compare against the experimental data. Initial tests confirm this is working as expected. Have updated run_tibd_pydream.py and the example in sim_protocols.py to use the new format.